Use struct initialization to ensure values are zero-initialized
In `seek_to.go`, we are unconditionally reading from `out.equal` but we are only
conditionally setting `out.equal` in `seek_to.c`. We could fix the bug by only
zeroing out the two fields that are unconditionally read, but zeroing out stack
memory is pretty cheap, so I'm erring on the safe side by initializing the whole
struct.