cgo_flags_common.go: remove -lpthread: Included by Cgo (#21)
This fixes the following build error:

    invalid flag in #cgo CXXFLAGS: -lpthread

This occurs because each of the Cgo compiler settings has a list of
acceptable flags, to try and make it harder to execute arbitrary
code when someone runs "go get". The -l flag is not permitted in
CXXFLAGS, but is in LDFLAGS.

cgo_snappy_common.go: cgo concatenates all #cgo directions in all .go
files. This means we were passing compiler flags multiple times.
Simplify the configuration to only pass flags once. This file had
many flags that were already specified in cgo_flags_common.go.